third_party/blink/web_tests/platform/generic/media/W3C/audio/error/error_null-expected.txt

audio.error - null if no source

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "null" is "null"
PASS successfullyParsed is true

TEST COMPLETE


spec reference
